Add deletion confirmation dialog'

+++ client/src/app/content-list/content-list.component.ts
@@ -1,12 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from "@angular/router";
-import { MatTableDataSource, MatSlideToggleChange } from '@angular/material';
+import { MatTableDataSource, MatSlideToggleChange, MatDialog } from '@angular/material';
 import { Subscription } from 'rxjs';
 import { ApiService } from '../api.service';
 import { Content } from '../content';
 import { User } from '../_models/user';
 import { AuthenticationService } from 'src/app/_services/authentication.service';
 import { AlertService } from 'src/app/_services/alert.service';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component'
 
 @Component({
   selector: 'app-content-list',
@@ -23,7 +24,8 @@ export class ContentListComponent implements OnInit {
   constructor(public apiService: ApiService,
     public authenticationService: AuthenticationService,
     private alertService: AlertService,
-    private router: Router) { }
+    private router: Router,
+    public dialog: MatDialog) { }
 
   ngOnInit() {
     const userJson = localStorage.getItem('currentUser');
@@ -40,8 +42,16 @@ export class ContentListComponent implements OnInit {
   }
 
   deleteContent = (id: number) => {
-    this.apiService.deleteContent(id).subscribe(() => {
-      window.location.reload();
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '350px',
+      data: "Do you confirm the deletion of this data?"
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        this.apiService.deleteContent(id).subscribe(() => {
+          window.location.reload();
+        });
+      }
     });
   }
